+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `package.json` entry point is `rollup:bundle` as to not coflict with the SDK's `es:bundle`. This is effectively not "turning on" this feature.
   But can still be opt-in. Generally, if this is needed, consumers should be utilizing the SDK's bundled es modules.
 * Converted `SessionTypes` to an `enum` but still left the union type as to not break consumers' typings.
-
+* Renamed `ISessionInfo` to `IPendingSession` (but kept `ISessionInfo` for backwards compatibility).
 # [v13.4.0](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v13.3.7...v13.4.0
 ### Added
 * [PCM-1753](https://inindca.atlassian.net/browse/PCM-1753) – Add an option for log formatters